fix(typescript): move Quasar CLI related typings to @quasar/app
Closes #6360
