+++ app/types/configuration.d.ts
@@ -1,0 +1,7 @@
+import { QuasarContext, QuasarConf } from "quasar";
+import "./configuration/conf";
+import "./configuration/context";
+
+declare module "quasar" {
+  type ConfigureCallback = (context: QuasarContext) => QuasarConf;
+}
