+++ app/types/configuration/boot.d.ts
@@ -1,0 +1,11 @@
+import "quasar";
+
+interface BootConfigurationItem {
+  path: string;
+  server?: false;
+  client?: false;
+}
+
+declare module "quasar" {  
+  type QuasarBootConfiguration = (string | BootConfigurationItem)[];
+}
