+++ app/types/configuration/capacitor-conf.d.ts
@@ -1,0 +1,16 @@
+import "quasar";
+
+declare module "quasar" {
+  type QuasarCapacitorTargets = "android" | "ios";
+  
+  interface QuasarCapacitorConfiguration {
+    /** If not present, will look for `package.json > capacitorId/cordovaId` */
+    id?: string;
+    /** If not present, will look for `package.json > name` */
+    appName?: string;
+    /** If not present, will look for `package.json > version` */
+    version?: string;
+    /** If not present, will look for `package.json > description` */
+    description?: string;
+  }
+}
