--- ui/types/configuration.d.ts
@@ -1,7 +1,0 @@
-import { QuasarConf } from "./configuration/conf";
-import { QuasarContext } from "./configuration/context";
-
-export * from "./configuration/conf";
-export * from "./configuration/context";
-
-export type ConfigureCallback = (context: QuasarContext) => QuasarConf;
