--- ui/types/configuration/boot.d.ts
@@ -1,7 +1,0 @@
-interface BootConfigurationItem {
-  path: string;
-  server?: false;
-  client?: false;
-}
-
-export type QuasarBootConfiguration = (string | BootConfigurationItem)[];
