+++ ui/types/globals.d.ts
@@ -1,8 +1,18 @@
 import { QuasarIconSet } from "./extras/icon-set";
 import { HasCapacitor, HasCordova, HasElectron } from "./feature-flag";
 import { QuasarLanguage } from "./lang";
-import * as Cordova from "cordova";
-import * as Electron from "electron";
+
+// We cannot reference directly Capacitor/Cordova/Electron types
+//  or they would generate TS errors for Vue CLI users
+// We also cannot move feature-flags system into `@quasar/app`
+//  because `QVueGlobals` augmentations won't be transfered to
+//  the `vue/types/vue` augmentation for unknown reason (probably TS limitations)
+//  and the system will just stop working
+// To workaround these problems we define an empty holder interface
+//  and augment it into `@quasar/app` with currect typings
+export interface GlobalsTypesHolder {
+  [index: string]: any;
+}
 
 interface GlobalQuasarLanguage extends QuasarLanguage {
   set(lang: QuasarLanguage): void;
@@ -16,8 +26,8 @@ interface GlobalQuasarIconSet extends QuasarIconSet {
 
 export interface QVueGlobals
   extends HasCapacitor<{ capacitor: any }>,
-    HasCordova<{ cordova: typeof Cordova }>,
-    HasElectron<{ electron: typeof Electron }> {
+    HasCordova<{ cordova: GlobalsTypesHolder["cordova"] }>,
+    HasElectron<{ electron: GlobalsTypesHolder["electron"] }> {
   version: string;
   lang: GlobalQuasarLanguage;
   iconSet: GlobalQuasarIconSet;
