+++ ui/wrappers/index.js
@@ -2,6 +2,9 @@
 //  they just take a callback function and return it
 // They're used to apply typings to the callback
 //  parameters and return value when using Quasar with TypeScript
+// We need these in `ui` folder to make `quasar/wrapper` import work,
+//  but they are useful only for Quasar CLI projects
+// They are typed via module augmentation by `@quasar/app`
 
 module.exports.boot = function (callback) {
   return callback
